refactor(cli): extract runtime config helpers
